requirements.txt
blinker==1.9.0
certifi==2025.7.9
click==8.2.1
cloudinary==1.44.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.13.2
python-dotenv==1.1.1
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
